requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
langgraph>=0.1.0
httpx>=0.27.0
trafilatura>=1.9.0
ddgs>=6.1.3
tenacity>=8.2.3
python-dotenv>=1.0.0
redis>=5.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
gspread>=5.12.0
openai>=1.0.0

# Enhanced UI dependencies
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
openpyxl>=3.1.0
